code cleaning
git-svn-id: http://svn.miranda-ng.org/main/trunk@11204 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/Scriver/src/globals.cpp
@@ -150,7 +150,7 @@ static int ackevent(WPARAM wParam, LPARAM lParam)
 		dbei.flags |= DBEF_UTF;
 	dbei.szModule = GetContactProto(hContact);
 	dbei.timestamp = time(NULL);
-	dbei.cbBlob = mir_strlen(item->sendBuffer) + 1;
+	dbei.cbBlob = (int)mir_strlen(item->sendBuffer) + 1;
 	if (!(item->flags & PREF_UTF))
 		dbei.cbBlob *= sizeof(TCHAR) + 1;
 	dbei.pBlob = (PBYTE)item->sendBuffer;
